pyproject.toml
[project]
name = "TMDB-Movie-Data-Fetching-Script"
version = "0.1.0"
description = "Simple script to fetch nearly all movie data from TMDB API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.1",
    "pandas>=2.2.3",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.6.8",
    "tqdm>=4.66.5",
]

[tool.ruff]
include = ["pyproject.toml"]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["G004", "D211", "D213"]